+++ cli/final.rb
@@ -1,6 +1,5 @@
 def export_json(data:)
   text = JSON.pretty_generate(data)
-  text.gsub!(/"\/\/ UNSET"/, "// UNSET")
 
   root = Pathname(__dir__).parent.parent
   route = root / ".squarkdown"